skills/community/clawhub/a/auditclaw-azure/scripts/requirements.txt
azure-identity==1.25.2
azure-mgmt-storage==24.0.0
azure-mgmt-network==30.2.0
azure-mgmt-compute==37.2.0
azure-mgmt-keyvault==13.0.0
azure-mgmt-sql==3.0.1
azure-mgmt-web==10.1.0
azure-mgmt-security==7.0.0
azure-mgmt-monitor==7.0.0
azure-mgmt-resource==25.0.0
